speechbrain model was replaced to its own file
